Handling the new datatype Fix some bugs after the handling

+++ src/components/ChangeButton.tsx
@@ -7,7 +7,7 @@ interface props {
   cardData: {
     front?: string;
     back?: string;
-    id?: BigInteger;
+    id?: bigint;
   };
 }
